app/requirements.txt
flask==2.0.3
meilisearch==0.18.3
gunicorn==20.1.0
requests==2.27.1
beautifulsoup4==4.11.1
PyJWT==2.4.0
Werkzeug==2.0.3
watchtower==3.0.0
boto3==1.20.53
urllib3==1.26.18
twilio==7.17.0 